refactor(backend): payments -> payment
