+++ packages/backend/src/index.ts
@@ -25,8 +25,8 @@ import { createPeerService } from './peer/service'
 import { createAccountService } from './open_payments/account/service'
 import { createSPSPRoutes } from './spsp/routes'
 import { createAccountRoutes } from './open_payments/account/routes'
-import { createIncomingPaymentRoutes } from './open_payments/payments/incoming/routes'
-import { createIncomingPaymentService } from './open_payments/payments/incoming/service'
+import { createIncomingPaymentRoutes } from './open_payments/payment/incoming/routes'
+import { createIncomingPaymentService } from './open_payments/payment/incoming/service'
 import { StreamServer } from '@interledger/stream-receiver'
 import { createWebhookService } from './webhook/service'
 import { createConnectorService } from './connector'
